refactor: better organization
