+++ src/pages/settings/shortcuts.tsx
