+++ src/pages/settings/sync/dropbox.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from "react";
-import { Note } from "../../store/types";
+import { Note } from "@/store/types";
 import { Browser } from "@capacitor/browser";
 import { SecureStoragePlugin } from "capacitor-secure-storage-plugin";
-import icons from "../../lib/remixicon-react";
+import icons from "@/lib/remixicon-react";
 const CLIENT_ID = import.meta.env.VITE_DROPBOX_CLIENT_ID;
 const CLIENT_SECRET = import.meta.env.VITE_DROPBOX_CLIENT_SECRET;
